Replaced the old contract version with the refactored one
