+++ .gitignore
@@ -1,7 +1,6 @@
-/node_modules
+**/node_modules
 /docs
 /build
 .DS_Store
 .env
-/client/node_modules
-/client/yarn-error.log
+**/yarn-error.log
